requirements.txt
# Python 3.13.7

fastapi==0.121.1
uvicorn[standard]==0.38.0
pydantic==2.12.4
python-dotenv==1.2.1
langgraph==1.0.3
langchain-ollama==1.0.0
langchain-openai==1.0.2
pytest==7.4.3